requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
openai==1.5.0
pydub==0.25.1
SpeechRecognition==3.10.0
sqlalchemy==2.0.23
pydantic==2.4.2
aiohttp==3.8.5
pytest==7.4.0
requests==2.31.0
google-generativeai==0.3.1
psycopg2-binary==2.9.9
asyncpg==0.28.0 